hoteltheme/requirements.txt
Django==2.2.5
django-debug-toolbar==2.1
django-filter==2.2.0
django-rest-passwordreset==1.0.0
django-uuidfield==0.5.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
virtualenv==16.7.5
